--- app/cho/transaction/operations/shared/validate.rb
@@ -1,22 +1,0 @@
-# frozen_string_literal: true
-
-module Transaction
-  module Operations
-    module Shared
-      class Validate
-        include Dry::Transaction::Operation
-
-        def call(change_set, additional_attributes: {})
-          change_set.validate(additional_attributes)
-          if change_set.valid?
-            Success(change_set)
-          else
-            Failure(change_set)
-          end
-        rescue StandardError => e
-          Failure(Transaction::Rejection.new('Error validating changeset', e))
-        end
-      end
-    end
-  end
-end
