+++ app/cho/transaction/shared/save_with_change_set.rb
@@ -6,13 +6,13 @@ module Transaction
       include Dry::Transaction(container: Operations::Container)
 
       # Operations will be resolved from the `Container` specified above
-      step :validate, with: 'shared.validate'
+      step :validate, with: 'change_set.validate'
       step :process_file, with: 'file.process'
       step :import_work, with: 'import.work'
       step :access_level, with: 'access_controls.access_level'
       step :system_creator, with: 'access_controls.system_creator'
       step :permissions, with: 'access_controls.permissions'
-      step :save, with: 'shared.save'
+      step :save, with: 'change_set.save'
     end
   end
 end
